+++ src/DigitList/InvalidDigitException.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace Riimu\Kit\BaseConversion\DigitList;
+
+/**
+ * Thrown whenever a value for invalid digit is requested.
+ * @author Riikka Kalliomäki <riikka.kalliomaki@gmail.com>
+ * @copyright Copyright (c) 2015, Riikka Kalliomäki
+ * @license http://opensource.org/licenses/mit-license.php MIT License
+ */
+class InvalidDigitException extends \InvalidArgumentException
+{
+
+}