+++ src/InvalidNumberBaseException.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace Riimu\Kit\BaseConversion;
+
+/**
+ * hrown if the number bases cannot be used by the conversion strategy.
+ * @author Riikka Kalliomäki <riikka.kalliomaki@gmail.com>
+ * @copyright Copyright (c) 2015, Riikka Kalliomäki
+ * @license http://opensource.org/licenses/mit-license.php MIT License
+ */
+class InvalidNumberBaseException extends \InvalidArgumentException
+{
+
+}